no more leaks

+++ execution/heredoc.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   heredoc.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mayyildi <mayyildi@student.42nice.fr>      +#+  +:+       +#+        */
+/*   By: sroggens <sroggens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/09 23:24:17 by sroggens          #+#    #+#             */
-/*   Updated: 2023/03/27 23:44:33 by mayyildi         ###   ########.fr       */
+/*   Updated: 2023/03/28 22:37:07 by sroggens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,11 @@
 void	countheredoc(t_list **lst)
 {
 	int	i;
+	int	j;
 	t_list	*tmp;
 
 	i = 0;
+	j = 0;
 	tmp = (*lst);
 	while (tmp)
 	{
@@ -25,11 +27,24 @@ void	countheredoc(t_list **lst)
 			i++;
 		tmp = tmp->next;
 	}
-	g_base.heredoc.countheredoc = 1;
+	tmp = (*lst);
+	while (tmp)
+	{
+		if (tmp->data == 1)
+			j++;
+		tmp = tmp->next;
+	}
+	g_base.heredoc.countheredoc = 0;
 	g_base.heredoc.totalheredoc = i;
+	g_base.heredoc.totalrealheredoc = j;
 	g_base.heredoc.processhere = 0;
-	g_base.heredoc.fdout = malloc(sizeof(int) * g_base.heredoc.totalheredoc);
-	g_base.heredoc.filename = malloc(sizeof(char *) * (g_base.heredoc.totalheredoc + 1));
+	g_base.heredoc.countalldoc = 0;
+	if (g_base.heredoc.totalheredoc > 0)
+	{
+		g_base.heredoc.fdout = malloc(sizeof(int) * g_base.heredoc.totalheredoc);
+		if (g_base.heredoc.totalrealheredoc > 0)
+			g_base.heredoc.filename = malloc(sizeof(char *) * (g_base.heredoc.totalrealheredoc + 1));
+	}
 }
 
 int	counthereinpipe(t_list **lst)
@@ -71,12 +86,14 @@ int	heredoc(t_list **lst)
 		{
 			line = readline("heredoc> ");
 			if (ft_strcmp(line, tmp->next->arg) == 0)
-				break ;
+				{
+					free(line);
+					break ;
+				}
 			if (g_base.sigint_received)
 				return (2) ;
 			if (line == NULL)
 				break ;
-			printf(RED "HERE" CRESET "\n");
 			ft_putstr_fd(line, g_base.heredoc.fdout[g_base.heredoc.countheredoc]);
 			ft_putstr_fd("\n", g_base.heredoc.fdout[g_base.heredoc.countheredoc]);
 			free(line);
@@ -88,17 +105,21 @@ int	heredoc(t_list **lst)
 		removenextnode(&tmp);
 		tmp->data = 11;
 		g_base.heredoc.countheredoc++;
+		g_base.heredoc.countalldoc++;
 		return (0);
 	}
 	else if (tmp->data == 2)
 	{
+		if (g_base.heredoc.countalldoc > 0)
+			close(g_base.heredoc.fdout[g_base.heredoc.countheredoc - 1]);
 		g_base.heredoc.fdout[g_base.heredoc.countheredoc] = open(tmp->next->arg, O_RDONLY);
 		tmp->next->data = 15;
 		tmp->data = 12;
-		g_base.heredoc.countheredoc++;
+		g_base.heredoc.countalldoc++;
 		return (0);
 	}
-	g_base.heredoc.filename[g_base.heredoc.countheredoc] = NULL;
+	if (g_base.heredoc.totalrealheredoc > 0)
+		g_base.heredoc.filename[g_base.heredoc.countheredoc] = NULL;
 	return (1);
 }
 
@@ -106,14 +127,15 @@ void	unlinkheredoc(void)
 {
 	int	i;
 
-	i = g_base.heredoc.countheredoc - 1;
-	while (i >= 1)
+	i = g_base.heredoc.totalrealheredoc;
+	if (i == 0)
+		return ;
+	while (i >= 0)
 		{
 			unlink(g_base.heredoc.filename[i]);
 			free(g_base.heredoc.filename[i]);
 			i--;
 		}
-
 	free(g_base.heredoc.filename);
 }
